# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Web Shop API
  description: API za upravljanje korisnicima, proizvodima i narudžbama
  version: 2.0.0
  contact:
    name: Web Shop API Support
    email: support@webshop.hr
servers:
  - url: http://localhost:8080
    description: Lokalni razvojni poslužitelj
...
paths:
...
  /users:
    get:
      operationId: getUsers
      summary: Dohvati sve korisnike
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Straničena lista korisnika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    get:
      operationId: getUserById
      summary: Dohvati korisnika po ID-u
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Korisnik pronađen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateUser
      summary: Ažuriraj korisnika
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Korisnik ažuriran
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteUser
      summary: Obriši korisnika
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Korisnik obrisan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/orders:
    get:
      operationId: getOrdersByUserId
      summary: Dohvati narudžbe korisnika
      tags:
        - Users
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Straničena lista narudžbi korisnika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── PRODUCTS ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token dobiven putem /auth/login endpointa

  # ── Reusable parameters ────────────────────────────────────────────────────
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Jedinstveni identifikator resursa
      schema:
        type: integer
        format: int64
        minimum: 1
    PageParam:
      name: page
      in: query
      description: Broj stranice (počinje od 0)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Broj elemenata po stranici
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: "Polje i smjer sortiranja (npr. createdAt,desc)"
      schema:
        type: string
        example: createdAt,desc

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Neispravan zahtjev — pogreška u validaciji podataka
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Korisnik nije autenticiran ili je token istekao
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Traženi resurs nije pronađen
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Adrian Brnić
        email:
          type: string
          format: email
          example: adrian.brnic@fer.hr
        phone:
          type: string
          example: "+38591000000"
        birthday:
          type: string
          format: date
          example: "2002-05-26"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-10T12:00:00Z"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Adrian Brnić
        phone:
          type: string
          pattern: '^\+?[0-9]{7,15}$'
          example: "+38591000000"
        birthday:
          type: string
          format: date
          example: "2002-05-26"
    UserPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # -- Product --
...
    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAID
        - SHIPPED
        - DELIVERED
        - CANCELLED
      example: CREATED
...
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        productId:
          type: integer
          format: int64
          example: 2
        productName:
          type: string
          example: Bežični miš
        unitPrice:
          type: number
          format: double
          example: 10.00
        quantity:
          type: integer
          example: 3
        subtotal:
          type: number
          format: double
          example: 30.00
...
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        totalPrice:
          type: number
          format: double
          example: 30.00
        status:
          $ref: '#/components/schemas/OrderStatus'
        createdAt:
          type: string
          format: date-time
          example: "2026-03-10T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-10T12:00:00Z"
    OrderPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # -- Shared --
    PageMetadata:
      type: object
      properties:
        number:
          type: integer
          description: Trenutna stranica (od 0)
          example: 0
        size:
          type: integer
          description: Broj elemenata po stranici
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Ukupan broj elemenata
          example: 42
        totalPages:
          type: integer
          description: Ukupan broj stranica
          example: 3
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-03-18T10:00:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Korisnik s ID-om 5 nije pronađen
        path:
          type: string
          example: /api/users/5